fix(refs T33737): use correct user for assignee check
There has been a rework of the token storage as well as the used user entity to move our application closer to Symfony standards. In that process, the token storage now returns a securityUser instead of the regular user entity. So we need to get the corresponding user entity via the UserFromSecurityUserProvider to restore the logic.