F t29585 adding addon controller paths (#1490)
* feat (refs T29585): adds controller_paths for addons

This new (optional) parameter in the addon-manifest allows addons to specify the locations of controllers so that the core can add the routes from annotations to the core router. This opens up the possibility of non-standard namespaces for controllers. Addons can either specifiy one path as a string or multiple paths as an array. If the parameter is not used, it defaults to the Symfony standard of `/src/Controller` which is the behaviour we used before. All addons should still work the same.

* fix (refs T29585): fixes controller_paths array handling

This also makes the fallback in AddonInfo.php obsolete as the config will always be an array.

* minor (refs T29585): simplify adding routes method

* fix (refs T29585): fixes loading of disabled addon routes

If an addon is not enabled, the routes from the addon should also not be available in the router.

* minor(refs T29585): removes unused const

---------

Co-authored-by: Sami Mussbach <mussbach@demos-deutschland.de>