fix(refs T29585): improves the default value handling
In this version you can either just ignore the value for 'controller_paths' in your addon manifest (like this: `controller_paths:`) or completely remove the key 'controller_paths' altogether. In both cases the system will default back to '/src/Controller' as per Symfony Best Practice. If you do want to specify one or more deviating paths, you need to add them as an array.

Sadly, I did not find any reasonable way to include the fallback for a single path as a string (i.e. `controller_paths: '/path/to/controllers'`) as that would require casting it as an array before normalization. But that messes with a default fallback for an empty key.