feat(refs T29585): adds default fallback for old installations
Since all of our addons so far have used the Symfony default controller path, I added it as a default fallback in case the 'controller_paths' key is missing in the manifest as it is for all previously installed addons.

This negates the need to reinstall all addons on all servers. And with time all addons will be reinstalled eventually anyway, at which point we can consider simply removing the default fallback as it should not be necessary anymore.