fix(refs T29585): improves the default value handling for addon controller_paths (#1862)
* fix(refs T29585): improves the default value handling

In this version you can either just ignore the value for 'controller_paths' in your addon manifest (like this: `controller_paths:`) or completely remove the key 'controller_paths' altogether. In both cases the system will default back to '/src/Controller' as per Symfony Best Practice. If you do want to specify one or more deviating paths, you need to add them as an array.

Sadly, I did not find any reasonable way to include the fallback for a single path as a string (i.e. `controller_paths: '/path/to/controllers'`) as that would require casting it as an array before normalization. But that messes with a default fallback for an empty key.

* feat(refs T29585): adds default fallback for old installations

Since all of our addons so far have used the Symfony default controller path, I added it as a default fallback in case the 'controller_paths' key is missing in the manifest as it is for all previously installed addons.

This negates the need to reinstall all addons on all servers. And with time all addons will be reinstalled eventually anyway, at which point we can consider simply removing the default fallback as it should not be necessary anymore.