feat: Adds dplanPermissions attribute (#1650)
* feat: adds attribute for plan permissions

This adds a new php attribute which acts exactly like the old dplanPermissions annotation. With the advent of attributes, we should replace the old annotation with the new, now better fitting, attribute.

* feat: adds attribute check to permission listener

This includes the new dplanPermisson attribute in permission listener to enable its usage. For the time being, both the annotation as well as the attribute can be used. But already the attribute would take precedent if it exists. This way, we can just delete the old annotation way when it is no longer needed as it acts as a kind of fallback already.

I also removed the old warning, that the annotation should be used as I don't think it is relevant anymore and it made the code much easier. There is also a logging in the permission processing when an empty array is handed in as permissions which would catch the same cases as before anyway.

* style: Apply php-cs-fixer

* style: Apply php-cs-fixer

* minor: adds class aliases

This should increase the readability in the code to better discern between attribute and annotation based permission.

* chore: corrects comment and log message

to better reflect that there is now also the opportunity to use attributes as well as annotation.

* minor: deprecates permissions annotation

as this should be replaced with attribute going forward

* Update demosplan/DemosPlanCoreBundle/Permissions/Permissions.php

Co-authored-by: breierDemos <106672028+breierDemos@users.noreply.github.com>

* Update demosplan/DemosPlanCoreBundle/Permissions/Permissions.php

Co-authored-by: breierDemos <106672028+breierDemos@users.noreply.github.com>

* Update demosplan/DemosPlanCoreBundle/Permissions/Permissions.php

Co-authored-by: breierDemos <106672028+breierDemos@users.noreply.github.com>

---------

Co-authored-by: Demos-CI <devops-system+github@demos-deutschland.de>
Co-authored-by: breierDemos <106672028+breierDemos@users.noreply.github.com>