Remove default margins for ul, ol elements
While inuitcss followed the (opinionated) approach of applying
a default margin to ul/ol elements, tailwind goes the other way
by removing margin for all elements by default. As we're designing
a webapp here, where ul elements are mostly used for semantic
purposes (not as part of content), the tailwind approach is the
way to go. Wherever we need margins on ol or ul, it will be applied
manually by using a spacing utility, or, if the intention is to
structure content for better readability, use a wrapper element with
the class c-styled-html.

So, the class "list-style-none" is removed from ul/ol elements, because
this is the new default, as the default margins are removed from inuitcss,
too.

See T27595.