feat(refs T32997 T31520): Implements SQL based Command to delete procedure
This is a rather risky and hacky solution for deleting a procedure with everything related to it. It is meant as an interim solution while we are working on a more stable way of deleting all related parts in the system preferably with Doctrine.

We need this solution to clean everything pertaining to customers that had to be deleted. Since we are not able to cleanly delete statements or procedures or some other related things via our repositories, this direct approach in SQL has been chosen to accomplish the task in a somewhat timely manner.