feat(refs T32997 T31520): wrapping everything in db transaction
Since we are interacting directly with the database, we have the problem that if something goes wrong later in the script, all changes until then will already be committed. This can create a rather unstable system. To make the command safer to use in that regard, all changes have been wrapped into a transaction.