Delete procedure and all related things command (via SQL) (#1958)
* feat(refs T32997 T31520): Implements SQL based Command to delete procedure

This is a rather risky and hacky solution for deleting a procedure with everything related to it. It is meant as an interim solution while we are working on a more stable way of deleting all related parts in the system preferably with Doctrine.

We need this solution to clean everything pertaining to customers that had to be deleted. Since we are not able to cleanly delete statements or procedures or some other related things via our repositories, this direct approach in SQL has been chosen to accomplish the task in a somewhat timely manner.

* style: Apply php-cs-fixer

* feat(refs T32997 T31520): adds tableExist checks

This way, the whole command should be a bit more robust in case there are changes in the db schema.

* feat(refs T32997 T31520): adds ES repopulate process

After doing all the changes to the database, we need to repopulate the ES to sync it back up with our current data. I added a command option to allow for easier debugging without triggering that, but it should only be used for that purpose!

* feat(refs T32997 T31520): wrapping everything in db transaction

Since we are interacting directly with the database, we have the problem that if something goes wrong later in the script, all changes until then will already be committed. This can create a rather unstable system. To make the command safer to use in that regard, all changes have been wrapped into a transaction.

* style: Apply php-cs-fixer

* minor(refs T32997 T31520): adds env handling

to make the ES repopulate actually usable on prod systems.

* minor(refs T32997 T31520): adds no-debug for ES

* feat(refs T32997 T31520): adds more tables to delete from

* fix(refs T32997 T31520): fixes IN string handling

To simply add an array of strings to the `setParameter` method, it also needs the correct type to correctly wrap it into quotes.

Also added some new deletions.

---------

Co-authored-by: Demos-CI <devops-system+github@demos-deutschland.de>