refactor: avoid entity wrapping for twig (#1897)
A previous attempt to pass objects into
twig files that use the same properties
as the resources returned by the generic
web API did not establish itself. The
attempt is thus rolled back,
simplifying code logic.

As the `ProcedureTypeResourceType` does
not apply any aliases or special
property permissions, no adjustment of
the twig files is necessary.