feat(refs T34626): adds addon wrapper to core template (#2005)
* feat(refs T34626): adds addon wrapper to core template

This is needed to allow addons to hook into this spot in the procedure settings in all projects. The explanation text is now also in the addons, so I only copied over the addon wrapper itself. I also changed the div class to match the ones around it. Lastly, the import of the AddonWrapper in the DpBasicSettings.vue was outdated and is now fixed.

* minor(refs T34626): removes redundant permission check

The template in the addon already checks for the correct permission to display the content, so we don't need it here. This has the added benefit that we don't need to preemptive 'know' the permission in the core.