fix(refs T19759): fix warnings/errors [vue 3 upgrade] (2322)
* refactor(refs T19759): remove productionTip

- config.productionTip has been removed in vue 3
- in prod build, the tip will not show up
see https://v3-migration.vuejs.org/breaking-changes/global-api.html#config-productiontip-removed

* refactor(refs T19759): define whitespace option

to remove a console warning, this option needs to be defined
see https://vuejs.org/api/application.html#app-config-compileroptions-whitespace

* refactor(refs T19759): replace Vue.set()

it has been removed in vue 3 and is no longer needed

* feat(refs T19759): add build feature flags

while these are set to the default values, defining them removes
a console warning:
vue.esm-bundler.js:6776 You are running the esm-bundler build of Vue.
It is recommended to configure your bundler to explicitly replace
feature flag globals with boolean literals to get proper tree-shaking
in the final bundle. See http://link.vuejs.org/feature-flags for more
details.

* feat(refs T19759): replace Vue.prototype

Vue.prototype is no longer available in vue 3. Instead,
app.config.globalProperties should be used

see https://v3-migration.vuejs.org/breaking-changes/global-api.html#a-new-global-api-createapp

* fix(refs T19759): use new hooks

in vue 3, the hooks in directives have changed:
bind -> beforeMount
inserted -> mounted
update -> updated
componentUpdated -> updated
unbind -> unmounted

* fix(refs T19759): replace this.$set

it has been removed in vue 3 and is no longer needed

* fix(refs T19759): remove unused attribute

SearchAndSorting doesn't have a `hasChangedStatements`
prop, so it was added as an attribute. I could not
find any usage of it, so I believe that it was added
here accidentally and can simply be removed

* fix(refs T19759): default to null

in vue 3, html attributes need to be set to null
instead of false to be removed

* fix(refs T19759): rename beforeDestroy to beforeUnmount

in vue 3, the beforeDestroy hook is replaced with
beforeUnmount

* build(refs T19759): update yarn.lock

* feat(refs T19759): use deep watcher

in vue 3, watchers for arrays are not
deep by default, so the option needs
to be set

#AB 9149
---------

Co-authored-by: Florian Salis <salis@demos-deutschland.de>
Co-authored-by: salisdemos <40487461+salisdemos@users.noreply.github.com>