fix(refs T35630): WIP add slider component
the behavior which triggers a bug may best be caught by guiding
users to not produce weird scale selections, which is best accomplished
with a range slider component. this way, users can only limit
the range edges.

vue-slider-component has a solid user base, gh stars, is actively maintained
and has a roadmap for vue3.