fix (refs T35735): add missing alias after rename of resource property
Previously a `type` attribute was exposed to the frontend for the
attachment resource type, which matched the `type` property in the
corresponding attachment entity. As `type` is defined as a reserved
name by the JSON:API specification and must not be used as attribute,
it was renamed to `attachmentType`. During the renaming the necessary
aliasing was forgotten, which is now added in this commit.