refactor (refs T35444): update EDT to current version (#2412)
* refactor (refs T35444): update EDT to current version

* style: Apply php-cs-fixer

* fix (refs T35735): add missing alias after rename of resource property

Previously a `type` attribute was exposed to the frontend for the
attachment resource type, which matched the `type` property in the
corresponding attachment entity. As `type` is defined as a reserved
name by the JSON:API specification and must not be used as attribute,
it was renamed to `attachmentType`. During the renaming the necessary
aliasing was forgotten, which is now added in this commit.

* fix (refs T35735): use correct constructor parameter

Previously a boolean was used to inform if there were
any properties that were not adjusted exactly as requested.
After a refactoring it now needs to be the list of those
properties.

* chore: update composer.lock after merge

* feat: T35735: Changed FE Handling from user/assignee from to-one to to-many

* chore: add missing type hint

* fix (refs T35751): always handle `logo` as relationship

Depending on the permission, `logo` was set as attribute,
which broke requests using it in an `include` field.

The added permission check is technically not necessary,
but improves clarity.

* fix (refs T35751): use ID instead of whole entity for comparison

Doctrine automatically uses the entity ID for comparisons,
the condition factory however expects a primitive value.

Using path instances allows static code checkers to detect
incomplete renamings.

* fix (refs T35758): inject correct logger instance

The parent class needs a more specific logger type
than the extending class provided

* style: Apply php-cs-fixer

* fix (refs T35735): update dplan with EDT patch

* fix (refs T35767 T35735): add EDT fix to dplan

* fix: request tag.title as its the proper attribute

* feat (refs T35444): Use type Claim for updating assignee in relationship to Statement

Relationship between Statement and assignee points to Claim resources but was updated until now with User resources. In the future when updating it the type of the relationship value to set must be set to Claim.

* feat (refs T35444): Fix request for orga list

- There are no resource types for currentSlug and statusInCustomer. CurrentSlug is an include and statusInCustomer is now OrgaStatusInCustomer.
- Empty page values are not allowed anymore. Add page 1 for initial request.
- No customer include for resource type Orga. We can access branding directly now.

* fix (refs T35799 T35792): use new EDT condition for inheritance check

* chore: update composer.lock after merge

* style: Apply php-cs-fixer

* fix: rename `GisLayer` resource attribute `type`

`type` is reserved by the JSON:API specification and
must not be used as attribute. This commit renames
the attribute in the backend. The frontend usage needs
to be adjusted in a follow-up commit.

* bug(refs T35767): Fix Gislayer attribute

Type is reserved and cannot be used as an attribute for resource type Gislayer. The new attribute is layerType.

* fix: rename `PriorityArea` resource attribute `type`

`type` is reserved by the JSON:API specification and
must not be used as attribute. This commit renames
the attribute in the backend. The frontend usage needs
to be adjusted in a follow-up commit.

* refactor: adjust core according to `demosplan-addon` dependency

Also remove addon-specific code.

* refactor: add methods into addon repository interfaces

Documentation should reside on the interface methods, and
thus was removed here.

* refactor: remove validation call

The validation of `segmentDraftList` was removed
in the main branch. This change just carries the
change over to the new code position.

* refactor (refs T35926): rename invalid resource type relationship

* style: Apply php-cs-fixer

* fix: add alias for non-symfony-project interface

* fix: add missing service interface alias

* fix: use working syntax for alias

* fix: use correct method to create resource type identifier

* refactor: resolve some phpstan concerns

* chore: update composer.lock after merge

* refactor: move method docblock into parent interface

* chore: simplify line syntax

* chore: update composer.lock after EDT dependency update

* chore: update composer.lock after merge

* chore: update composer.lock after changes in EDT dependency

* chore: update composer.lock after changes in EDT dependency

* chore: update composer.lock after changes in EDT dependency

* chore: update composer.lock after changes in EDT dependency

* chore: update composer.lock after changes in EDT dependency

* fix: use correct pagination type

The focused `PagePagination` instance needs to be
transformed into the more general `ApiPagination`
instance. Using `PagePagination` would be preferable,
but this would require refactorings of other methods
still calling `getEsFilteredObjects` with `ApiPagination`.

* chore: change dependency back to default for merge into main

* chore: update composer.lock after dependency changes

---------

Co-authored-by: Demos-CI <devops-system+github@demos-deutschland.de>
Co-authored-by: Florian Salis <salis@demos-deutschland.de>
Co-authored-by: Steffi <101879352+gruenbergerdemos@users.noreply.github.com>
Co-authored-by: Stefanie Grünberger <gruenberger@demos-deutschland.de>