chore: clean up moduleRules, use correct addons path in tw config
changes in config.cssPurge:
- rename cssPurge to purgeCss (that's how it is actually called)
- as we never set purgeCss.enabled to false, it can be removed
- if we rename the "paths" property to "content", the whole object
  can be spreaded into the options object
- the glob expander {} can be used to simplify content locations
- demosplan/**/ may contain html.twig, no "js.twig", vue, or js files as of now
- "tabs-component" is not used anymore
- "split-statement" is recognised as content source from addons,
  so can be removed from safelist

changes in moduleRules:
as we want to apply a different set of postCss plugins to postcss-loader
depending on the resourcePath of the file being processed (so that we can
skip purging the file generated by tailwindcss), we must provide two sets
of postCssPlugins. As these are used on multiple files, it did not work out
to just modify one set from within the postcssOptions callback, instead,
two objects had to be crafted beforehand.