Change build to support Tailwind in addons (#2593)
This change addresses the issue of being unable to utilize Tailwind 
within addon markup. The current limitation stems from the fact 
that addons lack their own style build and heavily depend on styles 
generated in demosplan-core. Two potential solutions were considered:

1. Incorporate a styles build into each addon, supporting Tailwind with 
our custom configuration and including the compiled style in demosplan projects.
2. find a way to build demosplan-core styles after an addon is installed, integrating 
the addon paths into the demosplan-core styles build logic. This requires tailwindcss
and purgeCss to be aware of these paths, ensuring classes used exclusively 
within addons are not inadvertently removed.

This pull request implements the second solution, which proved 
to be more feasible with minimal overhead. It introduces a new argument, 
`-S` (or `--only-styles`), to the frontend build command, enabling the 
exclusive building of styles for a faster process. Following addon installation, 
demosplan-core styles can now be built with this argument, incorporating 
the addon files into the tailwind config. As a result, Tailwind classes used 
in addons are included in the rendered styles.

Regarding some specific changes:

config.cssPurge:
- rename cssPurge to purgeCss (that's how it is actually called)
- as we never set purgeCss.enabled to false, it can be removed
- if we rename the "paths" property to "content", the whole object
  can be spreaded into the options object
- the glob expander {} can be used to simplify content locations
- demosplan/**/ may contain html.twig, no "js.twig", vue, or js files as of now
- "tabs-component" is not used anymore
- "split-statement" is recognised as content source from addons,
  so can be removed from safelist

moduleRules:
As we want to apply a different set of postCss plugins to postcss-loader
depending on the resourcePath of the file being processed (so that we can
skip purging the file generated by tailwindcss), we must provide two sets
of postCssPlugins. As these are used on multiple files, it did not work out
to just modify one set from within the postcssOptions callback, instead,
two objects had to be crafted beforehand.

Removal of CleanWebpackPlugin:
As of webpack 5.20.0 the output.clean option exists to
clean the output directory before emitting anything to it.
This enables us to only emit the style files with the
styles-only build, so that js resources are not deleted.

Also, we can remove CleanWebpackPlugin now.

---------

Co-authored-by: Stefan Graupner <graupner@demos-deutschland.de>
Co-authored-by: Demos-CI <devops-system+github@demos-deutschland.de>