fix (refs T30474): reduce inconsistent code duplication
Enhance the `OwnsProcedureConditionFactory` implementation, so
that it can not only create conditions to match unknown users
against a known procedure, but also to match unknown procedures
against a known user. Logically the comparison is the same in both
cases, but their implementation differs significantly.

Keeping the implementations as close together as possible avoids
divergences, like were already present in
`ProcedureService::getAdminProcedureConditions`. The matching
logic in this method behaved very similar to the one executed
via `ProcedureAccessEvaluator::isOwningProcedure` (which uses
`OwnsProcedureConditionFactory`), but was not exactly the same,
causing a bug. Hence, `isOwningProcedure` was selected as "truth" and
the logic in `getAdminProcedureConditions` was dropped.