fix (refs T30474): apply role checks correctly
When creating conditions to match procedures in the
database and thus evaluating user properties in PHP,
checking `isUserInCustomer` separately from the
role check does not work correctly if it is used with a
different customer than the current one.

Thus, when checking roles the customer must be provided
to avoid bugs when different customers than the current
one may be relevant in the future.

When creating conditions to match users in the database
(which means the roles and customer of the users are
evaluated in the database) this problem does not occur,
because the relational model and its evaluation will
match every condition against each row, so it does not
matter if the creation of those conditions happend
separately, as they are applied together in the database
anyway.