bug(refs T36258): fix slicing tagging (2675)
* bug(refs T36258): Add 'br' tag to allowed nodes for prosemirror

We also need to transform 'br' tags to prosemirror positions. Additionally exclude 'br' from markRegex, because otherwise it will be included due to 'b' tag and it will be transformed double (once as node and once as mark).

* bug(refs T36246): Allow char start and char end to be the same

While checking for overlapping segments we need to exclude the start and end char, because they can be the same for different segments.