fix (refs T30474): reduce inconsistent code duplication (2662)
* fix (refs T30474): reduce inconsistent code duplication

Enhance the `OwnsProcedureConditionFactory` implementation, so
that it can not only create conditions to match unknown users
against a known procedure, but also to match unknown procedures
against a known user. Logically the comparison is the same in both
cases, but their implementation differs significantly.

Keeping the implementations as close together as possible avoids
divergences, like were already present in
`ProcedureService::getAdminProcedureConditions`. The matching
logic in this method behaved very similar to the one executed
via `ProcedureAccessEvaluator::isOwningProcedure` (which uses
`OwnsProcedureConditionFactory`), but was not exactly the same,
causing a bug. Hence, `isOwningProcedure` was selected as "truth" and
the logic in `getAdminProcedureConditions` was dropped.

* fix (refs T30474): apply role checks correctly

When creating conditions to match procedures in the
database and thus evaluating user properties in PHP,
checking `isUserInCustomer` separately from the
role check does not work correctly if it is used with a
different customer than the current one.

Thus, when checking roles the customer must be provided
to avoid bugs when different customers than the current
one may be relevant in the future.

When creating conditions to match users in the database
(which means the roles and customer of the users are
evaluated in the database) this problem does not occur,
because the relational model and its evaluation will
match every condition against each row, so it does not
matter if the creation of those conditions happend
separately, as they are applied together in the database
anyway.

* style: Apply php-cs-fixer

* fix (refs T30474): do not authorize planning agency via config

* style: Apply php-cs-fixer

* refactor: rename method for better readability

* Update demosplan/DemosPlanCoreBundle/Logic/ProcedureAccessEvaluator.php

Co-authored-by: mussbach <57546580+mussbach@users.noreply.github.com>

---------

Co-authored-by: Demos-CI <devops-system+github@demos-deutschland.de>
Co-authored-by: mussbach <57546580+mussbach@users.noreply.github.com>