fix (refs T30474): adjust procedure "owning" evaluation
Previous refactorings tried to simplify unintuitive parts of the
evaluation, but introduced bugs due to misunderstandings of the
intended logic. This commit restores the original logic, while
going out of its way to avoid similar misunderstandings in the
future.

To be more specific: in the original, "correct" logic, variables
like `$orgaOwnsProcedure` did also extend the access for private
planning agencies if a global config parameter was enabled, that
had seemingly nothing to do with private planning agencies. The
new structure in this commit should make the different ways to
"own" a procedure more clear.