fix(refs T36055): correctly fallback to defaultMapExtent (2781)
When using DpOlMap without this._options.procedureExtent
being set to true, the default extent set in map_max_boundingbox
should be used. This extent has to be set via props.

As DpMapModal is always used within a procedure to
showcase drawings, we can use the extent
(Kartenbegrenzung) set for that specific procedure.