fix(refs T36655): fix all popover instances
as we removed the default configuration of v-popover
when handcrafting our own tooltip lib, some defaults
went abroad, such as the defaultTrigger and classes.
To not have to add the configuration in demosplan-ui
again, this can as well be solved via individual overrides
and adding the default selectors to our scss.