fix(refs T35941): rollback partial fix, serialize params
The fix applied in #2521 introduced the side effect that
the ref="recommendation" was undefined while the loading
state was applied to the EditableText instance, and so,
the `field` property was not correctly filled within the
payload of the entityTextSaved event. As the bug originates
in a change in dpApi, #2521 is partially rolled back.
Instead, the params in dpApi.get() are now serialized,
with the result that the code in
client/js/components/statement/assessmentTable/EditableText.vue:355
runs and everything is updated as intended.