feat: abort yarn install if demosplan-ui is mounted in container
As this would delete the local git repo when running "yarn install",
the "preinstall" hook cancels the installation. This is implemented
as a oneliner to not have to think about a good place for the bash script.
As we do not have the "docker" cli in the container, "cat proc/mounts" is
used to determine if the specific volume is mounted.