fix(refs DPLAN-11457): Remove "Dead" Code (2974)
The Backend Code was already moved to an addon.
This was supposed to happen with that logic too. But since its not in use anymore and would be really complicated to migrate to an addon, we decided to delete the FE Part.

Along I found some corresponding Templates without any usage.