feat: include preflight as separate file
The (slightly opinionated) tailwind browser reset,
"preflight", sets some defaults that we want to
override with the project css. However, as we also
want to include the tailwind utility classes last
(to get rid of making them !important), the preflight
styles need to be separately included, not within the
same file as the utility classes.

Other things changed when "driving by":
- PHPStorm seems to know @tailwind in current versions
- html.fixtinyfixed seems to no longer be used