fix: Adjust Type Casing for APi Types
we have to carefully check if the Types for API 1.0 Routes have to be Pascal or camel.
So do, some don't.
And with that and since to type doesn't get transformed anymore, we had to adjust the casing for `report` in the backend.
Along there where couple of ocurences, where changing to Pascal simply got overseen.