fix(refs  DPLAN-11960): WIP Make Layersort work again
with the changes to API 2.0 the patch-request and the data structure from the list request changed
Also the Permissions to update Layers and LayerCategories where not configured.

The FE now looks somehow ok. But I still get an error for the update-request.

I think when the BE works we have to adjust the FE again to then send all categories and layers one by on for the updated order position.