feat(refs DPLAN-12323, #AB15313): add data protection link to statement modal (3684)
* feat (DPLAN-12323) Add statementFormHintDataProtection field in UI and entity to store dataprotection text and make it flexible based on project

* feat (DPLAN-12323) Revert changes as data protection message is the same for all projects?

* feat (DPLAN-12323) Revert changes as data protection message is the same for all projects?

* feat (DPLAN-12323) Add translation and link for data protection text in translation general file

* feat (DPLAN-12323) Adjust translation so German text is on translation file

* feat (DPLAN-12323) Adjust translation so German text is on translation file

* feat(refs DPLAN-12323): override link color to make it accessible

* feat(refs DPLAN-12323): make icon align properly with text

* feat(refs DPLAN-12323): move text into existing info box

that way, there's only one info box and the error box is
displayed below it, not between two info boxes (which
might be confusing)

* feat(refs DPLAN-12323): move override to correct file

* feat(refs DPLAN-12323): remove unused class

* feat(refs DPLAN-12323, #AB15313): follow naming conventions for trans keys

the keys are named after the 3 steps in the statement modal, "write", "personal.data" and "recheck". we should follow that pattern here, too. Also, snake_case should be used since there is no hierarchy expressed, see https://dplan-documentation.demos-europe.eu/development/guidelines-conventions/best-practices/translation-keys.html#use-dots-and-snake-case

Co-authored-by: spiess-demos <40452344+spiess-demos@users.noreply.github.com>

* feat(refs DPLAN-12323, #AB15313): use new trans key after renaming

* feat(refs DPLAN-12323, #AB15313): use DpInlineNotification

- we have a component for this, let's use it
- instead of focussing the element containing the error message,
we now use aria-live="assertive". i'm not sure why the element
was focussed before, but it does not seem to make much sense, as
the element is not interactive and I believe it should suffice to
have the screen reader read the errors or highlight the notification
as well as the invalid form fields with red color

* style(refs DPLAN-12323, #AB15313): add bullet points to list of invalid fields

the invalid fields listed in the error message seemed kind of lost visually;
this was fixed by adding bullet points and aligning them with the text of
the error notification

* fix(refs DPLAN-12323): restore focussing of error notification

- the error notification was focussed so the errors will be read
out by a screen reader
- this may be implemented as well by setting aria-live="assertive",
but we are not sure if it breaks anything, so we decided to restore
the previous implementation

---------

Co-authored-by: Heidrun Wiemold <wiemold@demos-deutschland.de>
Co-authored-by: hwiem <40487668+hwiem@users.noreply.github.com>
Co-authored-by: spiess-demos <40452344+spiess-demos@users.noreply.github.com>