feat: adjust for addons:
Client Administration: Under Organizations, it should be possible to assign a MeinBerlinAddonOrganisationId. The ResourceType already exists.

FP-A: In the basic settings, there should be an additional field ProcedureShortName available with the add-on. The ResourceType already exists.

When the permission field_procedure_pictogram is enabled, the following fields should also be available in the core.