fix(refs DPLAN-3217): Double handling the Voters data to make it reactive
There is a bug in the vuex-json-api which leads to un-reactive behaviour especially when using the delete mutation.
To make it work anyway, we have to hold the voters data within the component too, so it remains reactive here. But we still need the propper state in the store to handle the api requests.

This only has to stay like so until we migrate to vue 3, which is hopefully in the near future