feat(refs DPLAN-12969, #AB21841): introduce store for FilterFlyout
to be able to easily access filter options in the FilterFlyout component,
they are moved to a store instead of passing them as props. that way,
when loading the grouped or ungrouped filter options after a user
opens a filter dropdown, we don't need to mutate the props, and it's
easier to understand what's going on. grouped and ungrouped
filter options can be written into the store from outside, then be
accessed directly in the FilterFlyout component. we should be able
to remove the data handling inside FilterFlyout now.