F necessary adjustments (3138)
* chore: add command information

* refactor: avoid exception for OpenApi generation

At some point the OpenApi generation should include
not only getable and listable types, but also ones
that can be created, deleted or updated.

To check what aspect a type supports, corresponding
methods are used. Until now those were set to throw an
exception by default, to give the developer a good hint
how to adjust the backend if support for that aspect is to
be added. However, those would hinder the OpenApi
generation. Using false instead is regression, the error
message when the client tries to access the type will just
be not as clear.

* fix: use correct property type hints

* style: Apply php-cs-fixer

* style: Apply php-cs-fixer

---------

Co-authored-by: Demos-CI <devops-system+github@demos-deutschland.de>
Co-authored-by: sierrazamudiodemos <152305675+sierrazamudiodemos@users.noreply.github.com>