fix: fix phpunit test after refactoring (2679)
* fix: fix phpunit test after refactoring

The concept of `PropertyDefinition`s and `IncludeDefinition`s
was previously improved on in the corresponding dependency and
these types deleted. This warrants a couple of adjustments
in the `DynamicTransformerTest`.

* style: Apply php-cs-fixer

* fix: use actual implementation to test type comparison

The readability- and transformer implementation will automatically
check if the PHP type of entity objects matches the expected one.
For this we need to inherit any kind of class/interface. For
this test `EntityInterface` is already the expected one, so implementing
it should be enough.

* fix: use correct class string

* style: Apply php-cs-fixer

* fix: add missing ID to test

The test data was once created with an empty ID,
because it was not needed. Now the ID is validated
in the engine and thus must no longer be empty in
the test.

* fix: assure matching IDs

* style: Apply php-cs-fixer

* Adjust test

* style: Apply php-cs-fixer

* Remove empty string so that does not fail

---------

Co-authored-by: Demos-CI <devops-system+github@demos-deutschland.de>
Co-authored-by: Sami Mussbach <mussbach@demos-deutschland.de>
Co-authored-by: Catalina Sierra Zamudio <sierra-zamudio@demos-deutschland.de>